ZABBIX_API_DPU/requirements.txt
python-dotenv>=0.19.0
requests>=2.25.0
psycopg2-binary>=2.9.0
pandas>=1.3.0